requirements.txt
streamlit
Pillow
pandas
numpy
imagehash
openpyxl
